refactor: internal/domain/types.goでの型定義を概念ごとにファイルを分割
