+++ split-pass-api/internal/domain/edge.go
@@ -1,0 +1,11 @@
+package domain
+
+// Edge は駅間データです
+type Edge struct {
+	FromID    int
+	ToID      int
+	EigyoKilo DeciKilo
+	GiseiKilo DeciKilo
+	IsLocal   bool
+	Company   CompanyID
+}
