+++ split-pass-api/internal/domain/station.go
@@ -1,0 +1,7 @@
+package domain
+
+// Station は駅データです
+type Station struct {
+	ID   int    `json:"id"`
+	Name string `json:"name"`
+}
